+++ backend/audit/intakelib/checks/check_eins_are_not_empty.py
@@ -1,6 +1,6 @@
 from django.core.exceptions import ValidationError
 import logging
-from .util import get_missing_value_errors
+from audit.intakelib.common import get_missing_value_errors
 
 logger = logging.getLogger(__name__)
