+++ backend/audit/intakelib/checks/check_federal_award_passed_through_optional.py
@@ -3,7 +3,7 @@ from audit.intakelib.intermediate_representation import (
     get_range_values_by_name,
     get_range_by_name,
 )
-from .util import get_message, build_cell_error_tuple
+from audit.intakelib.common import get_message, build_cell_error_tuple
 
 logger = logging.getLogger(__name__)
