+++ backend/audit/intakelib/checks/check_missing_required_fields.py
@@ -1,6 +1,6 @@
 from django.core.exceptions import ValidationError
 import logging
-from .util import get_missing_value_errors
+from audit.intakelib.common import get_missing_value_errors
 from audit.fixtures.excel import FORM_SECTIONS
 
 logger = logging.getLogger(__name__)
