+++ backend/audit/intakelib/checks/check_no_repeat_findings.py
@@ -1,6 +1,10 @@
 import logging
 from audit.intakelib.intermediate_representation import get_range_by_name
-from .util import get_message, build_cell_error_tuple, is_value_na
+from audit.intakelib.common import (
+    get_message,
+    build_cell_error_tuple,
+    is_value_marked_na,
+)
 
 logger = logging.getLogger(__name__)
 
@@ -13,7 +17,7 @@ def no_repeat_findings(ir):
     for ndx, (is_rep, prior) in enumerate(
         zip(repeat_prior_reference["values"], prior_references["values"])
     ):
-        if (is_rep == "N") and (not is_value_na(prior)):
+        if (is_rep == "N") and (not is_value_marked_na(prior)):
             errors.append(
                 build_cell_error_tuple(
                     ir,
@@ -22,7 +26,7 @@ def no_repeat_findings(ir):
                     get_message("check_no_repeat_findings_when_n"),
                 )
             )
-        elif (is_rep == "Y") and ((not prior) or (is_value_na(prior))):
+        elif (is_rep == "Y") and ((not prior) or (is_value_marked_na(prior))):
             errors.append(
                 build_cell_error_tuple(
                     ir,
