+++ backend/audit/intakelib/checks/check_other_cluster_names.py
@@ -1,6 +1,6 @@
 import logging
 from audit.intakelib.intermediate_representation import get_range_by_name
-from .util import get_message, build_cell_error_tuple
+from audit.intakelib.common import get_message, build_cell_error_tuple
 
 logger = logging.getLogger(__name__)
