+++ backend/audit/intakelib/checks/check_sequential_award_numbers.py
@@ -1,6 +1,6 @@
 import logging
 from audit.intakelib.intermediate_representation import get_range_by_name
-from .util import get_message, build_cell_error_tuple
+from audit.intakelib.common import get_message, build_cell_error_tuple
 import re
 
 logger = logging.getLogger(__name__)
