+++ backend/audit/intakelib/checks/check_uei_exists.py
@@ -1,6 +1,10 @@
 import logging
 from audit.intakelib.intermediate_representation import get_range_by_name
-from .util import list_contains_non_null_values, get_message, build_range_error_tuple
+from audit.intakelib.common import (
+    list_contains_non_null_values,
+    get_message,
+    build_range_error_tuple,
+)
 
 logger = logging.getLogger(__name__)
